common/config: enable section parsing when recreation
When RECREATE_TEST_DEV is enabled, specific fs mount options such as
TMPFS_MOUNT_OPTIONS are ignored during section parsing unless the FSTYP
is modified. This allows the filesystem type to remain the same while
only specific fs mount options are changed during the parsing of section
parameters for test device recreation or when the section with speciic
fs mount options are different from the default.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>