common/rc: print test mount options
Mount options for a SCRATCH device might not be the same for a TEST
device if RECREATE_TEST_DEV is not enabled. Add mount options for TEST
devices when printing fstest header to clarify this.

Add mount and mkfs info for TEST devices so we get the same information
being printed for both devices.

Export new TEST_{MKFS/MOUNT}_OPTIONS and include them in the report.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>