xfs/448: get rid of assert-on-failure
The bug this problem exercises has been fixed for quite some time,
but the test does not run on XFS debug kernels that have fatal
asserts enable. There is no reason for this now that the test does
not assert fail on most kernels regularly tested, so kill the
check and enable the test to run on all XFS configs.

Signed-off-by: Dave Chinner <dchinner@redhat.com>